// src/BindingChaos.Web/src/api/models/SignalDetailViewModel.ts
export interface SignalDetailViewModel {
    /**
     * The unique identifier of the signal.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    id: string;
    /**
     * The title of the signal.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    title?: string;
    /**
     * The description of the signal.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    description?: string;
    /**
     * The tags associated with the signal.
     * @type {Array<string>}
     * @memberof SignalDetailViewModel
     */
    tags?: Array<string>;
    /**
     * The pseudonym of the author.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    authorPseudonym?: string;
    /**
     * The creation timestamp in ISO 8601 format.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    createdAt?: string;
    /**
     * The last amplification timestamp in ISO 8601 format.
     * @type {string}
     * @memberof SignalDetailViewModel
     */
    lastAmplifiedAt?: string | null;
    /**
     * The total number of amplifications for this signal.
     * @type {number}
     * @memberof SignalDetailViewModel
     */
    amplifyCount?: number;
    /**
     * The list of amplifications for this signal.
     * @type {Array<AmplificationViewModel>}
     * @memberof SignalDetailViewModel
     */
    amplifications?: Array<AmplificationViewModel>;
    /**
     * Indicates whether the current user has amplified this signal.
     * @type {boolean}
     * @memberof SignalDetailViewModel
     */
    isAmplifiedByCurrentUser?: boolean;
    /**
     * Indicates whether the current user is the originator of this signal.
     * @type {boolean}
     * @memberof SignalDetailViewModel
     */
    isOriginator?: boolean;
    /**
     * The attachments associated with this signal.
     * @type {Array<AttachmentDetailViewModel>}
     * @memberof SignalDetailViewModel
     */
    attachments?: Array<AttachmentDetailViewModel>;
    /**
     * The suggested actions for this signal.
     * @type {Array<SuggestedActionViewModel>}
     * @memberof SignalDetailViewModel
     */
    suggestedActions?: Array<SuggestedActionViewModel>;
}