// src/BindingChaos.Web/src/api/models/ConcernOriginDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Describes how a concern came to be raised.
 * @export
 * @enum {string}
 */
export enum ConcernOriginDto {
    Manual = 'Manual',
    EmergingPattern = 'EmergingPattern'
}


export function instanceOfConcernOriginDto(value: any): boolean {
    for (const key in ConcernOriginDto) {
        if (Object.prototype.hasOwnProperty.call(ConcernOriginDto, key)) {
            if (ConcernOriginDto[key as keyof typeof ConcernOriginDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ConcernOriginDtoFromJSON(json: any): ConcernOriginDto {
    return ConcernOriginDtoFromJSONTyped(json, false);
}

export function ConcernOriginDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConcernOriginDto {
    return json as ConcernOriginDto;
}

export function ConcernOriginDtoToJSON(value?: ConcernOriginDto | null): any {
    return value as any;
}

export function ConcernOriginDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): ConcernOriginDto {
    return value as ConcernOriginDto;
}